+++ thoth-api/src/graphql/model.rs
@@ -1615,6 +1615,11 @@ impl MutationRoot {
         #[graphql(description = "Values for location to be created")] data: NewLocation,
     ) -> FieldResult<Location> {
         context.token.jwt.as_ref().ok_or(ThothError::Unauthorised)?;
+        // Only superusers can create new locations where Location Platform is Thoth
+        if !context.account_access.is_superuser && data.location_platform == LocationPlatform::Thoth
+        {
+            return Err(ThothError::ThothLocationError.into());
+        }
         context
             .account_access
             .can_edit(publisher_id_from_publication_id(
@@ -1970,12 +1975,33 @@ impl MutationRoot {
         #[graphql(description = "Values to apply to existing location")] data: PatchLocation,
     ) -> FieldResult<Location> {
         context.token.jwt.as_ref().ok_or(ThothError::Unauthorised)?;
-        let location = Location::from_id(&context.db, &data.location_id).unwrap();
+        let current_location = Location::from_id(&context.db, &data.location_id).unwrap();
+        let has_canonical_thoth_location = Publication::from_id(&context.db, &data.publication_id)?
+            .locations(
+                context,
+                Some(1),
+                None,
+                None,
+                Some(vec![LocationPlatform::Thoth]),
+            )?
+            .first()
+            .map_or(false, |location| location.canonical);
+        // Only superusers can update the canonical location when a Thoth Location Platform canonical location already exists
+        if has_canonical_thoth_location && data.canonical && !context.account_access.is_superuser {
+            return Err(ThothError::ThothUpdateCanonicalError.into());
+        }
+
+        // Only superusers can edit locations where Location Platform is Thoth
+        if !context.account_access.is_superuser
+            && current_location.location_platform == LocationPlatform::Thoth
+        {
+            return Err(ThothError::ThothLocationError.into());
+        }
         context
             .account_access
-            .can_edit(location.publisher_id(&context.db)?)?;
+            .can_edit(current_location.publisher_id(&context.db)?)?;
 
-        if data.publication_id != location.publication_id {
+        if data.publication_id != current_location.publication_id {
             context
                 .account_access
                 .can_edit(publisher_id_from_publication_id(
@@ -1989,7 +2015,7 @@ impl MutationRoot {
         }
 
         let account_id = context.token.jwt.as_ref().unwrap().account_id(&context.db);
-        location
+        current_location
             .update(&context.db, &data, &account_id)
             .map_err(|e| e.into())
     }
@@ -2294,6 +2320,12 @@ impl MutationRoot {
     ) -> FieldResult<Location> {
         context.token.jwt.as_ref().ok_or(ThothError::Unauthorised)?;
         let location = Location::from_id(&context.db, &location_id).unwrap();
+        // Only superusers can delete locations where Location Platform is Thoth
+        if !context.account_access.is_superuser
+            && location.location_platform == LocationPlatform::Thoth
+        {
+            return Err(ThothError::ThothLocationError.into());
+        }
         context
             .account_access
             .can_edit(location.publisher_id(&context.db)?)?;