+++ thoth-app/src/component/locations_form.rs
@@ -1,4 +1,5 @@
 use std::str::FromStr;
+use thoth_api::account::model::AccountDetails;
 use thoth_api::model::location::Location;
 use thoth_api::model::location::LocationPlatform;
 use thoth_errors::ThothError;
@@ -84,6 +85,7 @@ pub struct Props {
     pub locations: Option<Vec<Location>>,
     pub publication_id: Uuid,
     pub update_locations: Callback<()>,
+    pub current_user: AccountDetails,
 }
 
 impl Component for LocationsFormComponent {
@@ -145,7 +147,20 @@ impl Component for LocationsFormComponent {
                 {
                     FetchState::NotFetching(_) => vec![],
                     FetchState::Fetching(_) => vec![],
-                    FetchState::Fetched(body) => body.data.location_platforms.enum_values.clone(),
+                    FetchState::Fetched(body) => {
+                        if ctx.props().current_user.resource_access.is_superuser {
+                            body.data.location_platforms.enum_values.clone()
+                        // remove Thoth from LocationPlatform enum for non-superusers
+                        } else {
+                            body.data
+                                .location_platforms
+                                .enum_values
+                                .clone()
+                                .into_iter()
+                                .filter(|platform| platform.name != LocationPlatform::Thoth)
+                                .collect()
+                        }
+                    }
                     FetchState::Failed(_, _err) => vec![],
                 };
                 true
@@ -462,12 +477,27 @@ impl LocationsFormComponent {
             ctx.link()
                 .callback(move |_| Msg::DeleteLocation(location_id)),
         );
+        let mut edit_callback = Some(
+            ctx.link()
+                .callback(move |_| Msg::ToggleModalFormDisplay(true, Some(location.clone()))),
+        );
         let mut delete_deactivated = false;
+        let mut edit_deactivated = false;
+
         // If the location is canonical and other (non-canonical) locations exist, prevent it from
         // being deleted by deactivating the delete button and unsetting its callback attribute
         if l.canonical && ctx.props().locations.as_ref().unwrap_or(&vec![]).len() > 1 {
             delete_callback = None;
             delete_deactivated = true;
+        }
+        // If not superuser, restrict deleting and editing locations with Thoth location platform
+        if !ctx.props().current_user.resource_access.is_superuser
+            && l.location_platform == LocationPlatform::Thoth
+        {
+            delete_callback = None;
+            delete_deactivated = true;
+            edit_callback = None;
+            edit_deactivated = true;
         }
 
         html! {
@@ -510,7 +540,8 @@ impl LocationsFormComponent {
                         <div class="control">
                             <a
                                 class="button is-success"
-                                onclick={ ctx.link().callback(move |_| Msg::ToggleModalFormDisplay(true, Some(location.clone()))) }
+                                onclick={ edit_callback }
+                                disabled={ edit_deactivated }
                             >
                                 { EDIT_BUTTON }
                             </a>