+++ thoth-errors/src/lib.rs
@@ -103,6 +103,10 @@ pub enum ThothError {
     WithdrawnDateError,
     #[error("A Superseded or Withdrawn Work must have a Withdrawn Date.")]
     NoWithdrawnDateError,
+    #[error("Only superusers can create, edit, or delete Locations where the Location Platform is Thoth.")]
+    ThothLocationError,
+    #[error("Only superusers can update the canonical location when Thoth Location Platform is already set as canonical.")]
+    ThothUpdateCanonicalError,
 }
 
 #[cfg(not(target_arch = "wasm32"))]